// src/app/admin/pages/login-page/login-page.component.ts
import { Component, OnInit } from '@angular/core';
import { FormGroup, FormControl, Validators } from '@angular/forms';
import { User } from '../../shared/interfaces/user.interface';
import { AuthService } from '../../shared/services/auth.service';
import { Router, ActivatedRoute, Params } from '@angular/router';

@Component({
  selector: 'app-login-page',
  templateUrl: './login-page.component.html',
  styleUrls: ['./login-page.component.scss']
})
export class LoginPageComponent implements OnInit {
  form: FormGroup;
  message: string;
  constructor(
    public authS: AuthService,
    private router: Router,
    private route: ActivatedRoute
  ) { }

  ngOnInit(): void {
    this.route.queryParams.subscribe((params: Params) => {
      if (params['loginAgain']) {
        this.message = 'Введіть дані для входу'
      }
    });
    this.form = new FormGroup({
      email: new FormControl(null, [
        Validators.required, Validators.email
      ]),
      password: new FormControl(null, [
        Validators.required, Validators.minLength(6)
      ]),
    });
  }
  submitted: boolean;
  submit(): void{
    // console.log(this.form.controls.password.errors); //! errors
    
    this.submitted = true;
    if (this.form.invalid) {
      return
    }

    const user: User = {
      email: this.form.value.email,
      password: this.form.value.password
    }

    this.authS.login(user).subscribe((res) => { 
      this.form.reset();
      this.submitted = false;
      this.router.navigate(['/admin', 'dashboard']);
    }, (error) => {
        this.submitted = false;
    });
    
  }

}
